Add 'focus next' command
This command is similar to 'focus' command. The difference is that when
multiple windows matches the context, this command cycles between all
matched windows.

After writing this, I noticed that
http://i3wm.org/docs/hacking-howto.html#_using_git_sending_patches
mentions that you're not interested in alt-tab like window switching. If
you consider this alt-tab like, please just ignore this patch.