Add workspace match
This allows to match containers by workspace. The workspace can match
either by name or by it being current.

The rationale behind this is the following use case:

    bindsym $mod+q [class=XTerm workspace=current] focus

The match for current workspace is specified with the 'current' keyword,
which is not that nice, because this is also a valid regex. Ideas for
better syntax are welcome.

Currently, this is only added to the command parser. I'm not sure
whether it makes sense to add it also to the config file parser for use
in assign and for_window commands.