Implement a 'focus local|global <number>' directive.
This will focus the n-th child of either the focused container's parent (local) or the focused container's workspace (global).
Children are counted recursively, except for tabbed/stacked containers since a user only notices them as "one" window. If such
a tabbed or stacked container is selected, the currently focused container within will be focused instead of the split container.

fixes #1214