Extent tree_flatten to handle a second case.
If a v-split (h-split) container has another v-split (h-split) container
as its single child they can be collapsed to a single one.

This situation occurs when a container is unnecessarily split by
dragging a tiled container. The situation is difficult to detect in
advance, so we clean it up afterwards.