Simplify multi-monitor presentations
It seems that setting up i3 for multi-monitor presentations (e.g. with
LibreOffice Impress) is not as easy as it should be [1], [2]. I use the
method [1], but it is not ideal when one is in hurry or under stress
before the presentation.

This commit is my first attempt to solve this issue. I'm not sure wheter
this is the right way to implement this. Suggestions welcome :)

Programs like LO Impress or pdfpc configure the presentation windows to
have the same geometry as the outputs where the window should be. With
this change, when such a window is detected, it is open on the
appropriate workspace. Later, when the program sets its fullscreen
status, the window appear at the right output and the presentation works
out of the box.

Tested with LO Impress and pdfpc. The testsuite passes.

[1] https://faq.i3wm.org/question/2114/libreoffice-presentation-done-righttm-with-i3.1.html
[2] https://faq.i3wm.org/question/3965/dual-screen-presentation-does-not-work.1.html