Redraw right border correctly and only for non-pango fonts.
As the comment above the changed code correctly indicates, redrawing the
right client border is only necessary for non-pango fonts as there we have
no mechanism to ellipsize the text. We can be a little bit more efficient
by avoiding to draw in this case.

Furthermore, this patch replaces the x and width values with correctly
computed values. This prevents the redrawing to incorrectly cut off the
text of marks (or long titles).

relates to #2247