Indent container and window titles by the border width.
This patch indents window titles, and the representation used for
window-less containers, by the border width. This both looks visually more
appealing and is now consistent with the mark indentation from the previous
commit. For the typical border sizes of <= 3px, a user will hardly be able
to make out any difference, but when using bigger borders, the difference
becomes rather clear.

relates to #2247